6.3-stable

+++ share/mk/bsd.obj.mk
@@ -1,48 +1,65 @@
-#	$NetBSD: bsd.obj.mk,v 1.7 1995/06/10 20:46:35 mycroft Exp $
+#	$OpenBSD: bsd.obj.mk,v 1.18 2017/01/24 02:56:50 tb Exp $
+#	$NetBSD: bsd.obj.mk,v 1.9 1996/04/10 21:08:05 thorpej Exp $
 
 .if !target(obj)
-.if defined(NOOBJ)
+.  if defined(NOOBJ)
 obj:
-.else
+.  else
 
-.if defined(OBJMACHINE)
-__objdir=	obj.${MACHINE}
-.else
+.  if defined(MAKEOBJDIR)
+__objdir=	${MAKEOBJDIR}
+.  else
 __objdir=	obj
-.endif
+.  endif
 
-.if defined(USR_OBJMACHINE)
-__usrobjdir=	${BSDOBJDIR}.${MACHINE}
-__usrobjdirpf=	
-.else
-__usrobjdir=	${BSDOBJDIR}
-.if defined(OBJMACHINE)
-__usrobjdirpf=	.${MACHINE}
-.else
-__usrobjdirpf=
-.endif
-.endif
+_SUBDIRUSE:
 
-obj: _SUBDIRUSE
-	@cd ${.CURDIR}; rm -f ${__objdir} > /dev/null 2>&1 || true; \
-	here=`pwd`; subdir=`echo $$here | sed 's,^${BSDSRCDIR}/,,'`; \
-	if test $$here != $$subdir ; then \
-		dest=${__usrobjdir}/$$subdir${__usrobjdirpf} ; \
+obj! _SUBDIRUSE
+	@cd ${.CURDIR}; \
+	umask ${WOBJUMASK}; \
+	here=`/bin/pwd`; bsdsrcdir=`cd ${BSDSRCDIR}; /bin/pwd`; \
+	subdir=$${here#$${bsdsrcdir}/}; \
+	if [[ `id -u` -eq 0 && ${BUILDUSER} != root ]]; then \
+		SETOWNER="chown -h ${BUILDUSER}:${WOBJGROUP}"; \
+		if [[ $$here != $$subdir ]]; then \
+			_mkdirs() { \
+				su ${BUILDUSER} -c "mkdir -p $$1"; \
+			}; \
+			MKDIRS=_mkdirs; \
+		fi; \
+	elif [[ `id` == *'('${WOBJGROUP}')'* && $$here == $$subdir ]]; then \
+		SETOWNER="chown :${WOBJGROUP}"; \
+	else \
+		SETOWNER=:; \
+	fi; \
+	[[ -z $$MKDIRS ]] && MKDIRS="mkdir -p"; \
+	if [[ $$here != $$subdir ]]; then \
+		dest=${BSDOBJDIR}/$$subdir ; \
 		echo "$$here/${__objdir} -> $$dest"; \
-		rm -rf ${__objdir}; \
-		ln -s $$dest ${__objdir}; \
-		if test -d ${__usrobjdir} -a ! -d $$dest; then \
-			mkdir -p $$dest; \
+		if [[ ! -L ${__objdir} || `readlink ${__objdir}` != $$dest ]]; \
+		    then \
+			[[ -e ${__objdir} ]] && rm -rf ${__objdir}; \
+			ln -sf $$dest ${__objdir}; \
+			$$SETOWNER ${__objdir}; \
+		fi; \
+		if [[ -d ${BSDOBJDIR} ]]; then \
+			[[ -d $$dest ]] || $$MKDIRS $$dest; \
 		else \
-			true; \
+			if [[ -e ${BSDOBJDIR} ]]; then \
+				echo "${BSDOBJDIR} is not a directory"; \
+			else \
+				echo "${BSDOBJDIR} does not exist"; \
+			fi; \
 		fi; \
 	else \
-		true ; \
 		dest=$$here/${__objdir} ; \
-		if test ! -d ${__objdir} ; then \
+		if [[ ! -d ${__objdir} ]]; then \
 			echo "making $$dest" ; \
-			mkdir $$dest; \
+			$$MKDIRS $$dest; \
+			$$SETOWNER $$dest; \
 		fi ; \
 	fi;
+.  endif
 .endif
-.endif
+
+.include <bsd.own.mk>
